+++ beam-migrate/ChangeLog.md
@@ -11,6 +11,11 @@
     `foreignKeyConstraintSyntax`, for constructing foreign key constraint syntax.
   * Introduce `addTableForeignKey` for declaring new foreign key constraints.
 
+## Bug fixes
+
+* Fix an issue in which `beam-migrate` would fail to migrate a unique index
+  to a non-unique index or vice-versa.
+
 # 0.5.4.0
 
 ## Added features